requirements.txt
flask==2.3.0
tensorflow==2.13.0
numpy==1.24.3
reportlab==4.0.4
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.0
pillow==10.0.0


